requirements.txt
pyyaml
pycryptodome
construct